Add email verification function
